Fix bitsize of some msghdr and cmsghdr fields on Linux
These fields are size_t on Linux, so using CInt on LP64 only
accesses 32 bits of the total 64.  This is especially noticable on big
endian, where the lower 32 bits of the number are written to the upper
32 bits of the field.  Also use #type socklen_t instead of CUInt/CInt
where applicable, since POSIX does not specify the size of this type.

+++ include/HsNet.h
@@ -126,11 +126,11 @@ cmsg_nxthdr(struct msghdr *mhdr, struct cmsghdr *cmsg);
 extern unsigned char *
 cmsg_data(struct cmsghdr *cmsg);
 
-extern int
-cmsg_space(int l);
+extern size_t
+cmsg_space(size_t l);
 
-extern int
-cmsg_len(int l);
+extern size_t
+cmsg_len(size_t l);
 #endif /* _WIN32 */
 
 INLINE int
